+++ src/rhodes/__init__.py
@@ -1,4 +1,4 @@
 """Tools to create AWS Step Functions state machines."""
-__version__ = "0.2.0"
+__version__ = "0.2.1"
 
 from .states import StateMachine